+++ selene/support/shared/__init__.py
@@ -25,8 +25,8 @@ from __future__ import annotations
 from selene.support.shared.browser import SharedBrowser
 from selene.support.shared.config import SharedConfig
 
-from selene import managed
+from selene import _managed
 
 
-config = managed.config
-browser = managed.browser
+config = _managed.config
+browser = _managed.browser