+++ selene/support/shared/config.py
@@ -21,10 +21,10 @@
 # SOFTWARE.
 import warnings
 
-from selene import managed
+from selene.core.configuration import Config
 
 
-class SharedConfig(managed.Config):
+class SharedConfig(Config):
     warnings.warn(
-        'SharedConfig is deprecated. Use _Config instead', DeprecationWarning
+        'SharedConfig is deprecated. Use Config instead', DeprecationWarning
     )