+++ tests/acceptance/helpers/givenpage.py
@@ -20,14 +20,10 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-import os
+from tests import resources
 from tests.helpers import convert_sec_to_ms
 
-EMPTY_PAGE_URL = (
-    'file://'
-    + os.path.abspath(os.path.dirname(__file__))
-    + '/../../../resources/empty.html'
-)
+EMPTY_PAGE = resources.url('empty.html')
 
 
 class LoadingHtmlPage:
@@ -36,7 +32,7 @@ class LoadingHtmlPage:
         self._timeout = timeout
 
     def load_in(self, driver):
-        driver.get(EMPTY_PAGE_URL)
+        driver.get(EMPTY_PAGE)
         return LoadedHtmlPage(driver).render_body(self._body, self._timeout)